Incorrect return type for getAsyncString example
According to the documentation `-[BFTask continueWithBlock:]` returns a task. The `getAsyncString` example provided in the Readme.md makes it look like the string is returned synchronous, while in reality the creation of the string is still dependent on the async `getNumberAsync`, so a task need to be used to get the result.

To get the actual string you will still have to provide a continuation block on the returned task or wait for the block to finish first. The example code now makes it look like a string can magically be returned synchronously.